refactor(wave3): fix all remaining clippy warnings

+++ src/widgets/block.rs
@@ -12,7 +12,7 @@ use tui::{
 pub fn new<'a>(title: &'a str) -> Block<'a> {
     Block::default()
         .borders(Borders::ALL)
-        .border_style(Style::default().fg(Color::Indexed(239 as u8)))
+        .border_style(Style::default().fg(Color::Indexed(239_u8)))
         .title(title)
-        .title_style(Style::default().fg(Color::Indexed(249 as u8)))
+        .title_style(Style::default().fg(Color::Indexed(249_u8)))
 }
